All py in one file
