+++ src/config.json
@@ -1,1 +1,1 @@
-{"client_id": "", "client_secret": ""}
+{"client_id": "ikzSGBlERgDj0z5JfZ0WyA", "client_secret": "BSci9cTNUEcRR3LwlpbK07qhcg7CKQ"}
