+++ src/dat.json
@@ -1,1 +1,1 @@
-{"subreddit": "askreddit", "posts": 3, "comments": 0, "skipPosts": 0, "quality": 1, "shorts": false, "vidPath": "C:/Ajinkya/Code/vid.mp4"}
+{"subreddit": "askreddit", "posts": 3, "comments": 0, "skipPosts": 0, "quality": 1, "shorts": false, "vidPath": "null"}
